+++ Tutorials/01-BasicConcepts/01E-ActionSubscriber/ActionSubscriber/Store/EditCustomerUseCase/EditCustomerState.cs
@@ -1,0 +1,12 @@
+﻿namespace BasicConcepts.ActionSubscriber.Store.EditCustomerUseCase
+{
+	public class EditCustomerState
+	{
+		public bool IsLoading { get; private set; }
+
+		public EditCustomerState(bool isLoading)
+		{
+			IsLoading = isLoading;
+		}
+	}
+}