+++ Tutorials/01-BasicConcepts/01E-ActionSubscriber/ActionSubscriber/Store/EditCustomerUseCase/GetCustomerForEditAction.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace BasicConcepts.ActionSubscriber.Store.EditCustomerUseCase
+{
+	public class GetCustomerForEditAction
+	{
+		public Guid Id { get; }
+
+		public GetCustomerForEditAction(Guid id)
+		{
+			Id = id;
+		}
+	}
+}