+++ Tutorials/01-BasicConcepts/01E-ActionSubscriber/ActionSubscriber/Store/EditCustomerUseCase/GetCustomerForEditResultAction.cs
@@ -1,0 +1,14 @@
+﻿using BasicConcepts.ActionSubscriber.ApiObjects;
+
+namespace BasicConcepts.ActionSubscriber.Store.EditCustomerUseCase
+{
+	public class GetCustomerForEditResultAction
+	{
+		public CustomerEdit Customer { get; }
+
+		public GetCustomerForEditResultAction(CustomerEdit customer)
+		{
+			Customer = customer;
+		}
+	}
+}