+++ Tutorials/01-BasicConcepts/01E-ActionSubscriber/ActionSubscriber/Store/EditCustomerUseCase/Reducers.cs
@@ -1,0 +1,15 @@
+﻿using Fluxor;
+
+namespace BasicConcepts.ActionSubscriber.Store.EditCustomerUseCase
+{
+	public static class Reducers
+	{
+		[ReducerMethod]
+		public static EditCustomerState Reduce(EditCustomerState state, GetCustomerForEditAction action) =>
+			new EditCustomerState(isLoading: true);
+
+		[ReducerMethod]
+		public static EditCustomerState Reduce(EditCustomerState state, GetCustomerForEditResultAction action) =>
+			new EditCustomerState(isLoading: false);
+	}
+}